removed audio-ex sample because it is removed from the SDK
